fix(workbench): restore settings layer boundaries
Read the composer selection through an app-composition prop instead of
importing conversation state inside the settings feature, and move the
guide-to-settings target test into the guides layer that owns the contract.
